% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgppt-package.R
\name{geom_boxplot_jitter}
\alias{geom_boxplot_jitter}
\title{This geom is similar to \code{\link[ggplot2]{geom_boxplot}}, but allows to jitter outlier points and to raster points layer.}
\usage{
geom_boxplot_jitter(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
This geom is similar to \code{\link[ggplot2]{geom_boxplot}}, but allows to jitter outlier points and to raster points layer.
}
\section{Aesthetics}{



\code{geom_boxplot()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{lower}}
\item \strong{\code{upper}}
\item \strong{\code{middle}}
\item \strong{\code{ymin}}
\item \strong{\code{ymax}}
\item \code{alpha}
\item \code{colour}
\item \code{fill}
\item \code{group}
\item \code{linetype}
\item \code{shape}
\item \code{size}
\item \code{weight}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.


}

\examples{
ggplot() + geom_boxplot_jitter(aes(y=rt(1000, df=3), x=as.factor(1:1000 \%\% 2)), outlier.jitter.width = 0.1, raster = T)
}
