% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-utils.R
\name{geom_point_rast}
\alias{geom_point_rast}
\title{Raster version of geom_point}
\usage{
geom_point_rast(...)
}
\arguments{
\item{...}{arguments passed to ggrastr::geom_point_rast}
}
\description{
Raster version of geom_point
}
\seealso{
\code{\link[ggrastr]{geom_point_rast}}
}
