% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt-utils.R
\name{plot_base_ppt}
\alias{plot_base_ppt}
\title{Plot base R plots to ppt}
\usage{
plot_base_ppt(
  code,
  out_ppt,
  height = 6,
  width = 6,
  left = 5,
  top = 5,
  inches = FALSE,
  new_slide = FALSE,
  overwrite = FALSE,
  template = NULL,
  title = NULL,
  title_fp = NULL
)
}
\arguments{
\item{code}{code that generates plot}

\item{out_ppt}{output pptx file}

\item{height}{height in cm}

\item{width}{width in cm}

\item{left}{left alignment in cm}

\item{top}{top alignment in cm}

\item{inches}{measures are given in inches}

\item{new_slide}{add plot to a new slide}

\item{overwrite}{overwrite existing file}

\item{template}{path to a custom pptx template file. When NULL, uses
\code{getOption("tgppt.template")} if set, otherwise the bundled template.}

\item{title}{optional slide title string. When not NULL, a title text box is
placed above the plot.}

\item{title_fp}{an \code{fp_text} object controlling title font properties.
When NULL (the default), uses 18pt bold ArialMT.}
}
\value{
invisible; called for its side effect of writing a pptx file.
}
\description{
Plot base R plots to ppt
}
\examples{

temp_ppt <- tempfile(fileext = ".pptx")
plot_base_ppt(
    {
        plot(mtcars$mpg, mtcars$drat)
    },
    temp_ppt
)
}
