% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt-utils.R
\name{plot_base_ppt_at}
\alias{plot_base_ppt_at}
\title{Plot base R plots to ppt}
\usage{
plot_base_ppt_at(code, out_ppt, height = 6, width = 6, left = 5,
  top = 5, inches = FALSE, new_slide = FALSE)
}
\arguments{
\item{code}{code that generates plot}

\item{out_ppt}{output pptx file}

\item{height}{height in cm}

\item{width}{width in cm}

\item{left}{left alignment in cm}

\item{top}{top alignment in cm}

\item{inches}{measures are given in inches}

\item{new_slide}{add plot to a new slide}
}
\description{
Plot base R plots to ppt
}
\examples{

temp_ppt <- tempfile(fileext = ".pptx")
plot_base_ppt_at({plot(mtcars$mpg, mtcars$drat)}, temp_ppt)


}
