% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgppt-package.R
\docType{package}
\name{tgppt-package}
\alias{tgppt}
\alias{tgppt-package}
\title{tgppt: Tanay Group Powerpoint Utils}
\description{
Plot directly to Powerpoint. Based on the officer package.
}
\section{Options}{

\describe{
\item{\code{tgppt.template}}{Path to a custom PowerPoint template file (.pptx).
When set, this template is used by default in \code{\link[=new_ppt]{new_ppt()}}, \code{\link[=plot_gg_ppt]{plot_gg_ppt()}},
and \code{\link[=plot_base_ppt]{plot_base_ppt()}} instead of the bundled template. The \code{template}
parameter in those functions takes precedence over this option.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://tanaylab.github.io/tgppt/}
  \item \url{https://github.com/tanaylab/tgppt}
  \item Report bugs at \url{https://github.com/tanaylab/tgppt/issues}
}

}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

}
\keyword{internal}
