% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-utils.R
\name{vertical_labs}
\alias{vertical_labs}
\title{Set x-axis vertical labels in ggplot2}
\usage{
vertical_labs()
}
\value{
a \code{\link[ggplot2]{theme}} object that rotates x-axis labels to 90
degrees
}
\description{
Set x-axis vertical labels in ggplot2
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = factor(cyl), y = mpg)) +
    geom_boxplot() +
    vertical_labs()
}
