% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-utils.R
\name{arial_theme}
\alias{arial_theme}
\title{A ggplot2 custom theme (based on 'Arial' font)}
\usage{
arial_theme(size = 6)
}
\arguments{
\item{size}{base size of text}
}
\description{
A ggplot2 custom theme (based on 'Arial' font)
}
