% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-utils.R
\name{geom_boxplot_jitter}
\alias{geom_boxplot_jitter}
\title{Raster version of geom_boxplot with jittering of outliers}
\usage{
geom_boxplot_jitter(...)
}
\arguments{
\item{...}{arguments passed to ggrastr::geom_boxplot_jitter}
}
\description{
Raster version of geom_boxplot with jittering of outliers
}
\seealso{
\code{\link[ggrastr]{geom_boxplot_jitter}}
}
