% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt-utils.R
\name{plot_gg_ppt}
\alias{plot_gg_ppt}
\title{Plot ggplot object to ppt}
\usage{
plot_gg_ppt(
  gg,
  out_ppt,
  height = 6,
  width = 6,
  left = 5,
  top = 5,
  inches = FALSE,
  sep_legend = FALSE,
  transparent_bg = TRUE,
  rasterize_plot = FALSE,
  rasterize_legend = FALSE,
  res = 300,
  new_slide = FALSE,
  overwrite = FALSE,
  legend_height = NULL,
  legend_width = NULL,
  legend_left = NULL,
  legend_top = NULL
)
}
\arguments{
\item{gg}{ggplot object}

\item{out_ppt}{output pptx file}

\item{height}{height in cm}

\item{width}{width in cm}

\item{left}{left alignment in cm}

\item{top}{top alignment in cm}

\item{inches}{measures are given in inches}

\item{sep_legend}{plot legend and plot separatly}

\item{transparent_bg}{make background of the plot transparent}

\item{rasterize_plot}{rasterize the plotting panel}

\item{rasterize_legend}{rasterize the legend panel. Works only when sep_legend=TRUE and rasterize_plot=TRUE}

\item{res}{resolution of png used to generate rasterized plot}

\item{new_slide}{add plot to a new slide}

\item{overwrite}{overwrite existing file}

\item{legend_height, legend_width, legend_left, legend_top}{legend position and size in case \code{sep_legend=TRUE}. Similar to to heigh,width,left and top. By default - legend would be plotted to the right of the plot.}
}
\description{
Plot ggplot object to ppt
}
\examples{

library(ggplot2)
gg <- ggplot(mtcars, aes(x = mpg, y = drat)) +
    geom_point()
temp_ppt <- tempfile(fileext = ".pptx")
plot_gg_ppt(gg, temp_ppt)
}
