% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgppt-package.R
\docType{package}
\name{tgppt-package}
\alias{tgppt}
\alias{tgppt-package}
\title{tgppt: Tanay Group Powerpoint Utils}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

}
\keyword{internal}
