% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgppt-package.R
\docType{package}
\name{tgppt-package}
\alias{tgppt}
\alias{tgppt-package}
\title{tgppt: Tanay Group Powerpoint Utils}
\description{
Plot directly to Powerpoint. Based on the officer package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tanaylab/tgppt}
  \item Report bugs at \url{https://github.com/tanaylab/tgppt/issues}
}

}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

}
\keyword{internal}
