% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-utils.R
\name{theme_arial}
\alias{theme_arial}
\title{A ggplot2 custom theme (based on 'ArialMT' font)}
\usage{
theme_arial(size = 8)
}
\arguments{
\item{size}{base size of text}
}
\description{
A ggplot2 custom theme (based on 'ArialMT' font)
}
